+++ workflow_service/src/services/execute_action_service.py
@@ -1,12 +1,12 @@
 from src.database import async_session
 from src.logger import logger
 from src.rabbitmq.publisher import publish_action
-from core.get_chat_id import GetChatId
+from utils.get_chat_id import GetChatId
 from src.redis.redis_service import ServiceRedis
 from src.repositories.scenario_repo import ScenarioRepo
 from src.services.pattern.condition_service import ConditionService
 from src.services.pattern.event_service import EventService
-from core.get_user_email import GetUserEmail
+from utils.get_user_email import GetUserEmail
 from src.services.scenario_service import ScenarioService
